feat(trading limits): removed duplicated usage of sam estyled components
