+++ packages/blockchain-wallet-v4-frontend/src/modals/Settings/TradingLimits/template.success.gold.tsx
@@ -9,7 +9,7 @@ import Content from 'components/Flyout/Content'
 import { convertBaseToStandard } from 'data/components/exchange/services'
 import { LimitsAndDetails, SettingsLimit } from 'data/types'
 
-import { IconsContainer } from '../../components'
+import { IconsContainer, Title } from '../../components'
 import { Props as OwnProps, SuccessStateType } from '.'
 import { SETTINGS_ITEM_PERIOD, SETTINGS_ITEMS } from './model'
 import {
@@ -30,12 +30,7 @@ const Wrapper = styled.div`
   display: flex;
   flex-direction: column;
 `
-const Title = styled(Text)`
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin-bottom: 7px;
-`
+
 const CloseIconContainer = styled.div`
   display: flex;
   flex-direction: column;
@@ -126,7 +121,7 @@ const Template: React.FC<Props> = (props) => {
     <Wrapper>
       <HeaderWrapper>
         <IconsContainer>
-          <Title color='textBlack' size='24px' weight={600}>
+          <Title color='textBlack'>
             <FormattedMessage
               id='modals.onboarding.upgrade_now.account_limits'
               defaultMessage='Account Limits'
