+++ packages/blockchain-wallet-v4-frontend/src/modals/Settings/TradingLimits/template.success.silver.tsx
@@ -8,7 +8,7 @@ import { FlyoutWrapper } from 'components/Flyout'
 import Content from 'components/Flyout/Content'
 import { ModalName } from 'data/modals/types'
 
-import { IconsContainer } from '../../components'
+import { IconsContainer, Title } from '../../components'
 import { Props as OwnProps, SuccessStateType } from '.'
 import { TIER_TYPES } from './model'
 import {
@@ -31,12 +31,6 @@ const Wrapper = styled.div`
   height: 100%;
   display: flex;
   flex-direction: column;
-`
-const Title = styled(Text)`
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin-bottom: 7px;
 `
 
 const CloseIconContainer = styled.div`
@@ -90,7 +84,7 @@ const Template: React.FC<Props> = (props) => {
     <Wrapper>
       <HeaderWrapper>
         <IconsContainer>
-          <Title color='textBlack' size='24px' weight={600}>
+          <Title color='textBlack'>
             <FormattedMessage
               id='layouts.wallet.header.tradinglimits'
               defaultMessage='Trading Limits'
