+++ packages/blockchain-wallet-v4-frontend/src/modals/Settings/TradingLimits/template.success.tsx
@@ -11,7 +11,7 @@ import { convertBaseToStandard } from 'data/components/exchange/services'
 import { ModalName } from 'data/modals/types'
 import { SettingsItem, SettingsLimit } from 'data/types'
 
-import { IconsContainer } from '../../components'
+import { IconsContainer, Title } from '../../components'
 import { Props as OwnProps, SuccessStateType } from '.'
 import { SETTINGS_ITEM_PERIOD, SETTINGS_ITEMS, SETTINGS_ITEMS_ICONS, TIER_TYPES } from './model'
 import Gold from './template.success.gold'
@@ -29,12 +29,7 @@ const Wrapper = styled.div`
   display: flex;
   flex-direction: column;
 `
-const Title = styled(Text)`
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin-bottom: 7px;
-`
+
 const CloseIconContainer = styled.div`
   display: flex;
   flex-direction: column;
@@ -373,7 +368,7 @@ const Template: React.FC<Props> = (props) => {
             />
           </CloseIconContainer>
         </IconsContainer>
-        <Title color='textBlack' size='24px' weight={600} style={{ marginTop: '18px' }}>
+        <Title color='textBlack' style={{ marginTop: '18px' }}>
           <FormattedMessage
             id='modals.limits_and_features.title'
             defaultMessage='Limits & Features'
