+++ packages/blockchain-wallet-v4-frontend/src/modals/Settings/UpgradeNow/template.success.tsx
@@ -8,7 +8,7 @@ import { BlueCartridge } from 'components/Cartridge'
 import { FlyoutWrapper } from 'components/Flyout'
 import { ModalName } from 'data/types'
 
-import { IconsContainer } from '../../components'
+import { IconsContainer, Title } from '../../components'
 import { TIER_TYPES } from '../TradingLimits/model'
 import { Props as OwnProps, SuccessStateType } from '.'
 
@@ -18,12 +18,7 @@ const Wrapper = styled.div`
   display: flex;
   flex-direction: column;
 `
-const Title = styled(Text)`
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  margin-bottom: 7px;
-`
+
 const HeaderWrapper = styled(FlyoutWrapper)`
   position: fixed;
   max-width: 480px;
@@ -128,7 +123,7 @@ const Template: React.FC<Props> = (props) => {
             <Image width='26px' name='arrow-left' />
           </Link>
         </IconsContainer>
-        <Title color='textBlack' size='24px' weight={600} style={{ marginTop: '18px' }}>
+        <Title color='textBlack' style={{ marginTop: '18px' }}>
           <FormattedMessage id='scenes.interest.verifyid' defaultMessage='Upgrade Now' />
         </Title>
         <Text
